wip on refactoring the code to be cleaner, remove some deprecated calls/methods etc
