+++ src/main/java/uk/co/bitethebullet/android/token/dialogs/DeleteTokenPickerDialog.java
@@ -1,4 +1,4 @@
-package uk.co.bitethebullet.android.token;
+package uk.co.bitethebullet.android.token.dialogs;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -8,20 +8,24 @@ import android.os.Bundle;
 
 import androidx.fragment.app.DialogFragment;
 
-public class DeleteTokenDialog extends DialogFragment {
+import uk.co.bitethebullet.android.token.R;
+import uk.co.bitethebullet.android.token.dialogs.DeleteTokenDialog;
+
+//todo: MM add in radio button select list
+public class DeleteTokenPickerDialog extends DialogFragment {
 
     public interface DeleteTokenDialogListener {
         public void onDialogPositiveClick(DialogFragment dialog);
         public void onDialogNegativeClick(DialogFragment dialog);
     }
 
-    DeleteTokenDialogListener listener;
+    DeleteTokenDialog.DeleteTokenDialogListener listener;
 
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
         try {
-            listener = (DeleteTokenDialogListener) context;
+            listener = (DeleteTokenDialog.DeleteTokenDialogListener) context;
         } catch (ClassCastException e) {
             throw new ClassCastException("Must implement DeleteTokenDialogListener");
         }
@@ -32,8 +36,8 @@ public class DeleteTokenDialog extends DialogFragment {
 
         CharSequence tokenName = this.getArguments().getCharSequence("name");
         String message = getResources()
-                            .getString(R.string.delete_token_dialog_message)
-                            .replace("[name]", tokenName);
+                .getString(R.string.delete_token_dialog_message)
+                .replace("[name]", tokenName);
 
 
         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
@@ -41,12 +45,12 @@ public class DeleteTokenDialog extends DialogFragment {
                 .setTitle(R.string.delete_token_dialog_title)
                 .setPositiveButton(R.string.delete, new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
-                        listener.onDialogPositiveClick(DeleteTokenDialog.this);
+                        listener.onDialogPositiveClick(DeleteTokenPickerDialog.this);
                     }
                 })
                 .setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
-                        listener.onDialogNegativeClick(DeleteTokenDialog.this);
+                        listener.onDialogNegativeClick(DeleteTokenPickerDialog.this);
                     }
                 });
         // Create the AlertDialog object and return it
